vendor/info/mpix: Add MPIX_Info_dup_key().
This new function takes the handles of two info objects and copies
the value of the given `key` from the `srcinfo` to the `destinfo`
object. If the key is not found in `srcinfo`, `MPI_ERR_INFO_NOKEY`
is raised. If the key also existed in `destinfo`, the old value is
overwritten.

The purpose of this new function is to eliminate the need for the
application level to provide a buffer for the value itself during
this process. This is particularly useful when, for example, array
type entries are to be transferred from one info object to another
without having to deal with the values themselves.