rma: Implement RMA communication and synchronization.
This commit implements RMA communications using the pscom RMA
interfaces added with pscom-5.9.0-1 release. The functions to send and
receive RMA control messages and the callbacks to process
synchronization and datatype operations are also added.

This patch shifts the two-sided RMA handling to the pscom level as well
and unifies the RMA processing through this new interface.