vendor/ucc: Improve and extend datatype translation.
On the one hand, this commit adds support also for the newer MPI
standard datatypes such as `MPI_INT8_T` and family to the UCC
wrappers when doing the translation to the UCC datatypes.
In addition, also platform-dependent datatypes such as `MPI_INT`
are now mapped more accurately to their counterparts by checking
the actual sizes.

(Please note that this commit assumes that most of the code parts
added for making the related distinctions are consolidated during
compiler optimization due to constant folding and the subsequent
elimination of dead code.)