vendor/ucc: Add check for UCC-enabled device selection.
Even though the UCC support is basically independent of a specific
ADI3 device, it must also be used by the respective device in order
to take effect.

This commit therefore adds a check in the subconfigure of the UCC
wrappers that verifies whether the chosen device does this and aborts
if it does not. The UCC-enabled devices are listed explicitly in a
variable at the beginning of the script.