Fix configure check when testing for CUDA support in UCC.
The first argument for AC_LANG_PROGRAM is missing an additional layer
of braces, while the second argument defines a main which is automatically
added by AC_LANG_PROGRAM.

This fails the test with NVHPC and other Clang-based compilers, and might
produce unexpected results with other compilers.

To fix this, apply proper syntax to the ucc subconfigure.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>