test/errhan: Remove invalid info test case from sessionargs test.
This test case is not compliant with the MPI ABI and hence we
remove it from the sessionargs test. In particular, for the MPI
ABI there is no check if a value passed as MPI info is actually
an MPI info handle or not. So such an error cannot be detected.