vendor/ucc: Adapt to new vci param in progress hook.
With upstream 8bd29c0 (i.e., vendor 4.3.2), a new vci parameter
has been added to the `MPIR_Progress_hook_register()` function.
This commit is to adapt to this interface change. However, as
only the `MPIR_Async` hook will make use of this new parameter,
we can simply stick to the case of global progress (`vci=-1`).