vendor/ucc: Make priority/timing of ucc finalize more flexible.
The timing of the `ucc_finalize()` call may depend on the device, i.e.,
whether UCC should still be active in `MPID_Finalize()` or not.
This commit introduces `MPIDI_COMMON_UCC_FINALIZE_CALLBACK_PRIO` as a
new macro that can be set/overwritten by the respective device and that
determines the point in time of the `mpidi_ucc_finalize()` invocation
in the series of finalize callbacks. The default is that it is called
right _before_ `MPID_Finalize()`.