vendor/ucc: Release ucc team for comm_world during finalize.
Since the point in time of `ucc_finalize()` can now be changed by the
device using the new `MPIDI_COMMON_UCC_FINALIZE_CALLBACK_PRIO` macro,
it is possible that the corresponding finalize callback can be invoked
before the destroy hook for `MPI_COMM_WORLD` is called.
This commit ensures that the `ucc_team` for `MPI_COMM_WORLD` is always
released before UCC is finalized.